Updates to fix import namespace
