+++ src/fusion_tools/__init__.py
@@ -1,5 +1,3 @@
 
 from fusion_tools.visualization import Visualization
-from fusion_tools.handler import *
-
